Additional of FreeRTOS SDK (#51)
* vt middleware for freertos

* fix cmake

* cleanup

* structure changes

* first push

* log update

* vt_fc_read fix & reset fingerprint crash fix

* middleware update

* Common VT library (#49)

* formatting update

* test cases

* Doxygen comments

* Break up CMake across multiple files, update readme and doxygen

* formating fix

* fix tests

* refactor API documentation

* rework resource requirements

* various changes based on comments.

* define update

* add support markdown

* remove unused header, use strtok instead of strtok_r for compatibility

+++ src/core/fallcurve/internal/vt_fc_signature_compute.c
@@ -119,7 +119,7 @@ VT_UINT fc_signature_compute(
     VT_INT32 decimal    = pearson_coeff_computed;
     VT_FLOAT frac_float = pearson_coeff_computed - (VT_FLOAT)decimal;
     VT_INT32 frac       = frac_float * 10000;
-    VTLogDebug("Pearson Coeff: %lu.%lu \r\n", decimal, frac);
+    VTLogDebug("Pearson Coeff: %lu.%lu \r\n", (long unsigned int)decimal, (long unsigned int)frac);
 #endif /* VT_LOG_LEVEL > 2 */
 
     if (sample_length > VT_FC_MIN_FALLTIME_DATAPOINTS && pearson_coeff_computed > VT_FC_MIN_SHAPE_MATCH)
